% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureExtraction_covariateData.R
\name{ATCgroups}
\alias{ATCgroups}
\title{ATCgroups}
\usage{
ATCgroups(
  connection,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  cdmDatabaseSchema,
  cdmVersion = "5",
  cohortTable = "#cohort_person",
  cohortIds = c(-1),
  rowIdField = "subject_id",
  covariateSettings,
  aggregated = FALSE,
  minCharacterizationMean = 0,
  ...
)
}
\arguments{
\item{connection}{A database connection object created using \code{DatabaseConnector::connect}.}

\item{tempEmulationSchema}{The temp schema where the covariate tables will be created.}

\item{cdmDatabaseSchema}{The schema where the cdm tables are located.}

\item{cdmVersion}{The version of the cdm.}

\item{cohortTable}{The table where the cohort data is located.}

\item{cohortIds}{The cohort ids to include.}

\item{rowIdField}{The field in the cohort table that is the row id.}

\item{covariateSettings}{A list of settings for the covariate data.}

\item{aggregated}{Logical. If TRUE, the covariate data is aggregated.}

\item{minCharacterizationMean}{The minimum mean for the covariate to be included.}

\item{...}{Additional arguments (not currently used).}
}
\description{
ATCgroups
}
