% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDMdbHandler.R
\name{CDMdbHandler}
\alias{CDMdbHandler}
\title{CDM Database Handler Class}
\description{
A class for handling CDM database connections and operations
}
\section{Methods}{

\describe{
\item{\code{initialize(databaseId)}}{Initialize a new CDM database handler}
\item{\code{loadConnection(loadConnectionChecksLevel)}}{Load database connection with specified check level}
}
}

\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{databaseId}}{The ID of the database}

\item{\code{databaseName}}{The name of the database}

\item{\code{databaseDescription}}{Description of the database}

\item{\code{connectionHandler}}{Handler for database connections}

\item{\code{vocabularyDatabaseSchema}}{Schema name for the vocabulary database}

\item{\code{cdmDatabaseSchema}}{Schema name for the CDM database}

\item{\code{resultsDatabaseSchema}}{Schema name for the results database}

\item{\code{connectionStatusLog}}{Log of connection status and operations}

\item{\code{vocabularyInfo}}{Information about the vocabulary tables}

\item{\code{CDMInfo}}{Information about the CDM structure}

\item{\code{getTblVocabularySchema}}{Function to get vocabulary schema table}

\item{\code{getTblCDMSchema}}{Function to get CDM schema table}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CDMdbHandler-new}{\code{CDMdbHandler$new()}}
\item \href{#method-CDMdbHandler-loadConnection}{\code{CDMdbHandler$loadConnection()}}
\item \href{#method-CDMdbHandler-closeConnection}{\code{CDMdbHandler$closeConnection()}}
\item \href{#method-CDMdbHandler-clone}{\code{CDMdbHandler$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CDMdbHandler-new"></a>}}
\if{latex}{\out{\hypertarget{method-CDMdbHandler-new}{}}}
\subsection{Method \code{new()}}{
Returns the results database schema name.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CDMdbHandler$new(
  databaseId,
  databaseName,
  databaseDescription,
  connectionHandler,
  cdmDatabaseSchema,
  vocabularyDatabaseSchema = cdmDatabaseSchema,
  resultsDatabaseSchema = cdmDatabaseSchema,
  loadConnectionChecksLevel = "allChecks"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{databaseId}}{A text id for the database the it connects to}

\item{\code{databaseName}}{A text name for the database the it connects to}

\item{\code{databaseDescription}}{A text description for the database the it connects to}

\item{\code{connectionHandler}}{A ConnectionHandler object}

\item{\code{cdmDatabaseSchema}}{Name of the CDM database schema}

\item{\code{vocabularyDatabaseSchema}}{(Optional) Name of the vocabulary database schema (default is cdmDatabaseSchema)}

\item{\code{resultsDatabaseSchema}}{(Optional) Name of the results database schema (default is cdmDatabaseSchema)}

\item{\code{loadConnectionChecksLevel}}{(Optional) Level of checks to perform when loading the connection (default is "allChecks")
Reload connection}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CDMdbHandler-loadConnection"></a>}}
\if{latex}{\out{\hypertarget{method-CDMdbHandler-loadConnection}{}}}
\subsection{Method \code{loadConnection()}}{
Updates the connection status by checking the database connection, vocabulary database schema, and CDM database schema.
Close Connection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CDMdbHandler$loadConnection(loadConnectionChecksLevel)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{loadConnectionChecksLevel}}{Level of connection checks to perform}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CDMdbHandler-closeConnection"></a>}}
\if{latex}{\out{\hypertarget{method-CDMdbHandler-closeConnection}{}}}
\subsection{Method \code{closeConnection()}}{
Explicitly closes the database connection. This method should be called
when you're done with the CDMdbHandler to properly clean up resources.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CDMdbHandler$closeConnection()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CDMdbHandler-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CDMdbHandler-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CDMdbHandler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
