% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortData.R
\name{checkCohortData}
\alias{checkCohortData}
\alias{assertCohortData}
\title{Check if a tibble is of cohortData format.}
\usage{
checkCohortData(tibble)

assertCohortData(tibble)
}
\arguments{
\item{tibble}{The tibble to be checked}
}
\value{
TRUE if the tibble is of cohortData format, an array of strings with the failed checks
}
\description{
This function performs various validations on the cohortData tibble to ensure its integrity and correctness. The validations include:
}
\details{
\itemize{
\item Checking if cohortData is a tibble.
\item Checking if the cohortData tibble contains the required column names: cohort_name, person_source_value, cohort_start_date, cohort_end_date.
\item Validating the types of cohort_name, person_source_value, cohort_start_date, and cohort_end_date columns (character and Date types, respectively).
\item Checking for missing values in cohort_name and person_source_value columns.
\item Verifying the order of cohort_start_date and cohort_end_date, ensuring cohort_start_date is older than cohort_end_date.
}
}
