% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortTable.R
\name{checkCohortDefinitionTables}
\alias{checkCohortDefinitionTables}
\title{Check Existence of Cohort Definition Tables}
\usage{
checkCohortDefinitionTables(
  connectionDetails = NULL,
  connection = NULL,
  cohortDatabaseSchema,
  cohortDefinitionTable = "cohort_definition",
  cohortTable = "cohort"
)
}
\arguments{
\item{connectionDetails}{Details required to establish a database connection (optional).}

\item{connection}{An existing database connection (optional).}

\item{cohortDatabaseSchema}{The database schema where the cohort definition tables are expected to be found.}

\item{cohortDefinitionTable}{The name of the cohort definition table. Defaults to "cohort_definition".}

\item{cohortTable}{The name of the table containing cohort data. Defaults to "cohort".}
}
\value{
A tibble log indicating the success or error in checking the existence of tables.
}
\description{
Validates the existence of the specified cohort definition and cohort tables within a given database schema.
Requires either an active database connection or connection details to create one. This function is intended
to assist in ensuring that the necessary tables for cohort definition are present in the database.
}
