% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortsSummary.R
\name{checkCohortsSummary}
\alias{checkCohortsSummary}
\alias{assertCohortsSummary}
\title{Check if a tibble is of CohortsSummary format.}
\usage{
checkCohortsSummary(tibble)

assertCohortsSummary(tibble)
}
\arguments{
\item{tibble}{The tibble to be checked}
}
\value{
TRUE if the tibble is of CohortsSummary format, an array of strings with the failed checks
}
\description{
This function performs various validations on the CohortsSummary tibble to ensure its integrity and correctness. The validations include:
}
\details{
\itemize{
\item Checking if CohortsSummary is a tibble.
\item Checking if the CohortsSummary tibble contains the required column names
\item Validating the types
}
}
