% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_DatabasesSettings.R
\name{checkDatabaseSettings}
\alias{checkDatabaseSettings}
\title{checkDatabaseSettings}
\usage{
checkDatabaseSettings(database_settings)
}
\arguments{
\item{database_settings}{list with database settings}
}
\value{
tibble with check resutls
}
\description{
checkDatabaseSettings
}
