% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortData.R
\name{cohortDataToCohortDefinitionSet}
\alias{cohortDataToCohortDefinitionSet}
\title{cohortDataToCohortDefinitionSet}
\usage{
cohortDataToCohortDefinitionSet(
  cohortData,
  newCohortIds = NULL,
  skipCohortDataCheck = FALSE
)
}
\arguments{
\item{cohortData}{The tibble in cohortData format.}

\item{newCohortIds}{A vector of new cohort IDs to be assigned to the cohort data.}

\item{skipCohortDataCheck}{Logical value indicating whether to skip cohort data validation (default is FALSE).}
}
\value{
A data frame of format cohortDefinitionSet.
}
\description{
Convert a cohortData tibble into a cohortDefinitionSet
}
