% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortDiagnostics.R
\name{CohortDiagnostics_mergeCsvResults}
\alias{CohortDiagnostics_mergeCsvResults}
\title{CohortDiagnostics_mergeCsvResults}
\usage{
CohortDiagnostics_mergeCsvResults(
  pathToResultFolders,
  pathToMergedRestulsFolder
)
}
\arguments{
\item{pathToResultFolders}{character vector of paths to the result folders}

\item{pathToMergedRestulsFolder}{path to the folder where the merged results will be stored}
}
\value{
path to the merged results folder
}
\description{
Merge the results of multiple CohortDiagnostics into a single folder
}
