% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortDiagnostics.R
\name{CohortDiagnostics_runTimeCodeWAS}
\alias{CohortDiagnostics_runTimeCodeWAS}
\title{CohortDiagnostics_runTimeCodeWAS}
\usage{
CohortDiagnostics_runTimeCodeWAS(
  connectionDetails = NULL,
  connection = NULL,
  cdmDatabaseSchema,
  vocabularyDatabaseSchema = cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable = "cohort",
  cohortIdCases,
  cohortIdControls,
  covariateSettings = FeatureExtraction::createDefaultTemporalCovariateSettings()
)
}
\arguments{
\item{connectionDetails}{A list of details needed to create a database connection, or a pre-existing database connection object.}

\item{connection}{A pre-existing database connection object.}

\item{cdmDatabaseSchema}{The name of the schema containing the Common Data Model (CDM) tables.}

\item{vocabularyDatabaseSchema}{The name of the schema containing the vocabulary tables.}

\item{cohortDatabaseSchema}{The name of the schema containing the cohort table.}

\item{cohortTable}{The name of the cohort table.}

\item{cohortIdCases}{The ID of the cohort representing cases.}

\item{cohortIdControls}{The ID of the cohort representing controls.}

\item{covariateSettings}{A list of settings for extracting temporal covariates.}
}
\value{
A data frame with the counts of cases and controls for each time window and covariate, as well as the results of a Fisher's exact test comparing the counts of cases and controls.
}
\description{
This function calculates the number of subjects with observation case and controls in each time window for a given cohort. It returns a data frame with the counts of cases and controls for each time window and covariate, as well as the results of a Fisher's exact test comparing the counts of cases and controls.
}
