% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortGenerator.R
\name{CohortGenerator_deleteCohortFromCohortTable}
\alias{CohortGenerator_deleteCohortFromCohortTable}
\title{Delete cohort from cohort table}
\usage{
CohortGenerator_deleteCohortFromCohortTable(
  connectionDetails = NULL,
  connection = NULL,
  cohortDatabaseSchema,
  cohortTableNames,
  cohortIds
)
}
\arguments{
\item{connectionDetails}{A list containing the necessary details to establish a database connection.}

\item{schema}{The name of the schema where the cohort table and cohort names table are located.}

\item{cohort_table_name}{The name of the cohort table.}

\item{cohort_names}{A character vector of cohort names to be deleted.}
}
\value{
TRUE if the deletion is successful.
}
\description{
This function deletes specified cohorts from the cohort table and cohort names table in the database.
}
