% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortGenerator.R
\name{CohortGenerator_dropCohortStatsTables}
\alias{CohortGenerator_dropCohortStatsTables}
\title{CohortGenerator_dropCohortStatsTables}
\usage{
CohortGenerator_dropCohortStatsTables(
  connectionDetails = NULL,
  connection = NULL,
  cohortDatabaseSchema,
  cohortTableNames = CohortGenerator::getCohortTableNames()
)
}
\arguments{
\item{connectionDetails}{An object of class \code{connectionDetails} containing database connection details.}

\item{connection}{A database connection object created using \code{DatabaseConnector::connect}.}

\item{cohortDatabaseSchema}{The schema where the cohort tables will be created.}

\item{cohortTableNames}{A list containing the names of the cohort tables to be created.}
}
\description{
Wrapper for CohortGenerator::dropCohortStatsTables, where if bigquery is used, the tables are deleted using bigrquery::bq_table_delete.
}
