% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortGenerator.R
\name{CohortGenerator_getCohortDemograpics}
\alias{CohortGenerator_getCohortDemograpics}
\title{Get cohort counts and basic cohort demographics for a specific cohort.}
\usage{
CohortGenerator_getCohortDemograpics(
  connectionDetails = NULL,
  connection = NULL,
  cdmDatabaseSchema,
  vocabularyDatabaseSchema = cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable = "cohort",
  cohortIds = c(),
  toGet = c("histogramCohortStartYear", "histogramCohortEndYear", "histogramBirthYear",
    "sexCounts"),
  cohortDefinitionSet = NULL,
  databaseId = NULL
)
}
\arguments{
\item{connectionDetails}{Details required to establish a database connection (optional).}

\item{connection}{An existing database connection (optional).}

\item{cdmDatabaseSchema}{The schema name for the Common Data Model (CDM) database.}

\item{vocabularyDatabaseSchema}{The schema name for the vocabulary database (default is \code{cdmDatabaseSchema}).}

\item{cohortDatabaseSchema}{The schema name for the cohort database.}

\item{cohortTable}{The name of the cohort table in the cohort database (default is "cohort").}

\item{cohortIds}{A numeric vector of cohort IDs for which to retrieve demographics (default is an empty vector).}

\item{toGet}{A character vector indicating which demographic information to retrieve. Possible values include "histogramCohortStartYear", "histogramCohortEndYear", "histogramBirthYear", and "sexCounts".}

\item{cohortDefinitionSet}{A set of cohort definitions (optional).}

\item{databaseId}{The ID of the database (optional).}
}
\value{
A data frame with cohort counts and selected demographics
}
\description{
This function retrieves demographic information for a given cohort .
}
