% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortGenerator.R
\name{CohortGenerator_getCohortsOverlaps}
\alias{CohortGenerator_getCohortsOverlaps}
\title{CohortGenerator_getCohortsOverlaps}
\usage{
CohortGenerator_getCohortsOverlaps(
  connectionDetails = NULL,
  connection = NULL,
  cohortDatabaseSchema,
  cohortTable = "cohort",
  cohortIds = c()
)
}
\arguments{
\item{connectionDetails}{A list containing details for connecting to the database. Default is NULL.}

\item{connection}{A database connection object. Default is NULL.}

\item{cohortDatabaseSchema}{The schema where the cohort table resides.}

\item{cohortTable}{The name of the cohort table.}

\item{cohortIds}{A vector of cohort ids for which to calculate overlaps. Default is an empty vector.}
}
\value{
A tibble containing the overlaps between cohorts.
With one logical column for each cohort id and a column \code{numberOfSubjects} with the number of subjects in the combination.
}
\description{
Description: A function to calculate overlaps between cohorts.
}
