% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortTableHandler.R
\name{CohortTableHandler}
\alias{CohortTableHandler}
\title{CohortTableHandler}
\description{
Class for handling cohort tables in a CDM database.
Inherits from CDMdbHandler.
}
\section{Super class}{
\code{\link[HadesExtras:CDMdbHandler]{HadesExtras::CDMdbHandler}} -> \code{CohortTableHandler}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{cohortDatabaseSchema}}{Schema where cohort tables are stored}

\item{\code{cohortTableNames}}{Names of the cohort tables in the database}

\item{\code{cohortDefinitionSet}}{Set of cohort definitions}

\item{\code{cohortGeneratorResults}}{Results from cohort generation process}

\item{\code{cohortDemograpics}}{Demographic information for cohorts}

\item{\code{cohortsOverlap}}{Information about overlapping cohorts}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CohortTableHandler-new}{\code{CohortTableHandler$new()}}
\item \href{#method-CohortTableHandler-loadConnection}{\code{CohortTableHandler$loadConnection()}}
\item \href{#method-CohortTableHandler-insertOrUpdateCohorts}{\code{CohortTableHandler$insertOrUpdateCohorts()}}
\item \href{#method-CohortTableHandler-deleteCohorts}{\code{CohortTableHandler$deleteCohorts()}}
\item \href{#method-CohortTableHandler-getCohortCounts}{\code{CohortTableHandler$getCohortCounts()}}
\item \href{#method-CohortTableHandler-getNumberOfSubjects}{\code{CohortTableHandler$getNumberOfSubjects()}}
\item \href{#method-CohortTableHandler-getNumberOfCohortEntries}{\code{CohortTableHandler$getNumberOfCohortEntries()}}
\item \href{#method-CohortTableHandler-getCohortsSummary}{\code{CohortTableHandler$getCohortsSummary()}}
\item \href{#method-CohortTableHandler-getCohortIdAndNames}{\code{CohortTableHandler$getCohortIdAndNames()}}
\item \href{#method-CohortTableHandler-updateCohortNames}{\code{CohortTableHandler$updateCohortNames()}}
\item \href{#method-CohortTableHandler-getCohortsOverlap}{\code{CohortTableHandler$getCohortsOverlap()}}
\item \href{#method-CohortTableHandler-getNumberOfOverlappingSubjects}{\code{CohortTableHandler$getNumberOfOverlappingSubjects()}}
\item \href{#method-CohortTableHandler-getSexFisherTest}{\code{CohortTableHandler$getSexFisherTest()}}
\item \href{#method-CohortTableHandler-getYearOfBirthTests}{\code{CohortTableHandler$getYearOfBirthTests()}}
\item \href{#method-CohortTableHandler-closeConnection}{\code{CohortTableHandler$closeConnection()}}
\item \href{#method-CohortTableHandler-clone}{\code{CohortTableHandler$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortTableHandler-new"></a>}}
\if{latex}{\out{\hypertarget{method-CohortTableHandler-new}{}}}
\subsection{Method \code{new()}}{
Initialize the CohortTableHandler object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortTableHandler$new(
  connectionHandler,
  databaseId,
  databaseName,
  databaseDescription,
  cdmDatabaseSchema,
  vocabularyDatabaseSchema = cdmDatabaseSchema,
  resultsDatabaseSchema = cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTableName,
  loadConnectionChecksLevel = "allChecks"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{connectionHandler}}{The connection handler object.}

\item{\code{databaseId}}{A text id for the database the it connects to}

\item{\code{databaseName}}{A text name for the database the it connects to}

\item{\code{databaseDescription}}{A text description for the database the it connects to}

\item{\code{cdmDatabaseSchema}}{Name of the CDM database schema.}

\item{\code{vocabularyDatabaseSchema}}{Name of the vocabulary database schema. Default is the same as the CDM database schema.}

\item{\code{resultsDatabaseSchema}}{(Optional) Name of the results database schema (default is cdmDatabaseSchema).}

\item{\code{cohortDatabaseSchema}}{Name of the cohort database schema.}

\item{\code{cohortTableName}}{Name of the cohort table.}

\item{\code{loadConnectionChecksLevel}}{(Optional) Level of checks to perform when loading the connection (default is "allChecks")

loadConnection}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortTableHandler-loadConnection"></a>}}
\if{latex}{\out{\hypertarget{method-CohortTableHandler-loadConnection}{}}}
\subsection{Method \code{loadConnection()}}{
Reloads the connection with the initial setting and updates connection status

insertOrUpdateCohorts
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortTableHandler$loadConnection(loadConnectionChecksLevel)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{loadConnectionChecksLevel}}{Level of checks to perform during connection}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortTableHandler-insertOrUpdateCohorts"></a>}}
\if{latex}{\out{\hypertarget{method-CohortTableHandler-insertOrUpdateCohorts}{}}}
\subsection{Method \code{insertOrUpdateCohorts()}}{
If there is no cohort with the same cohortId it is added to the cohortDefinitionSet,
If there is a cohort with the same cohortId, the cohort is updated in the cohortDefinitionSet
CohortDefinitionSet is generated and demographics is updated for only the cohorts that have changed
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortTableHandler$insertOrUpdateCohorts(cohortDefinitionSet)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cohortDefinitionSet}}{The cohort definition set to add.

deleteCohorts}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortTableHandler-deleteCohorts"></a>}}
\if{latex}{\out{\hypertarget{method-CohortTableHandler-deleteCohorts}{}}}
\subsection{Method \code{deleteCohorts()}}{
Deletes cohorts from the cohort table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortTableHandler$deleteCohorts(cohortIds)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cohortIds}}{The cohort ids to delete.

getCohortCounts}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortTableHandler-getCohortCounts"></a>}}
\if{latex}{\out{\hypertarget{method-CohortTableHandler-getCohortCounts}{}}}
\subsection{Method \code{getCohortCounts()}}{
Retrieves cohort counts from the cohort table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortTableHandler$getCohortCounts()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A tibble containing the cohort counts with names.

getNumberOfSubjects
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortTableHandler-getNumberOfSubjects"></a>}}
\if{latex}{\out{\hypertarget{method-CohortTableHandler-getNumberOfSubjects}{}}}
\subsection{Method \code{getNumberOfSubjects()}}{
Retrieves the number of subjects of a cohort ID.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortTableHandler$getNumberOfSubjects(selected_cohortId)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{selected_cohortId}}{The cohort id for which number of subjects is sought.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A numeric value of the number of subjects in a cohort.

getNumberOfCohortEntries
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortTableHandler-getNumberOfCohortEntries"></a>}}
\if{latex}{\out{\hypertarget{method-CohortTableHandler-getNumberOfCohortEntries}{}}}
\subsection{Method \code{getNumberOfCohortEntries()}}{
Retrieves the number of entries of a cohort ID.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortTableHandler$getNumberOfCohortEntries(selected_cohortId)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{selected_cohortId}}{The cohort id for which number of entries is sought.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A numeric value of the number of entries in a cohort.

getCohortsSummary
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortTableHandler-getCohortsSummary"></a>}}
\if{latex}{\out{\hypertarget{method-CohortTableHandler-getCohortsSummary}{}}}
\subsection{Method \code{getCohortsSummary()}}{
Retrieves the summary of cohorts including cohort start and end year histograms and sex counts.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortTableHandler$getCohortsSummary(includeAllEvents = F)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{includeAllEvents}}{Logical, whether to include all events or just subjects. Default is FALSE.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A tibble containing cohort summary.

getCohortIdAndNames
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortTableHandler-getCohortIdAndNames"></a>}}
\if{latex}{\out{\hypertarget{method-CohortTableHandler-getCohortIdAndNames}{}}}
\subsection{Method \code{getCohortIdAndNames()}}{
Retrieves the cohort names.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortTableHandler$getCohortIdAndNames()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A vector with the name of the cohorts

updateCohortNames
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortTableHandler-updateCohortNames"></a>}}
\if{latex}{\out{\hypertarget{method-CohortTableHandler-updateCohortNames}{}}}
\subsection{Method \code{updateCohortNames()}}{
Updates the cohort name and short name.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortTableHandler$updateCohortNames(cohortId, newCohortName, newShortName)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cohortId}}{The cohort id to update.}

\item{\code{newCohortName}}{New name to assign to the cohort}

\item{\code{newShortName}}{New short name to assign to the cohort}

\item{\code{cohortName}}{The new cohort name.}

\item{\code{shortName}}{The new short name.

getCohortsOverlap}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortTableHandler-getCohortsOverlap"></a>}}
\if{latex}{\out{\hypertarget{method-CohortTableHandler-getCohortsOverlap}{}}}
\subsection{Method \code{getCohortsOverlap()}}{
Retrieves the number of subjects that are in more than one cohort.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortTableHandler$getCohortsOverlap()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A tibble containing one logical column for each cohort with name a cohort id,
and an additional column \code{numberOfSubjects} with the number of subjects in the cohorts combination.

getNumberOfOverlappingSubjects
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortTableHandler-getNumberOfOverlappingSubjects"></a>}}
\if{latex}{\out{\hypertarget{method-CohortTableHandler-getNumberOfOverlappingSubjects}{}}}
\subsection{Method \code{getNumberOfOverlappingSubjects()}}{
Retrieves the number of subjects that are overlapping between two given cohorts.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortTableHandler$getNumberOfOverlappingSubjects(
  selected_cohortId1,
  selected_cohortId2
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{selected_cohortId1}}{The cohort id of the first cohort.}

\item{\code{selected_cohortId2}}{The cohort id of the second cohort.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A numeric value of the number of overlapping subjects between two given cohorts.

getSexFisherTest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortTableHandler-getSexFisherTest"></a>}}
\if{latex}{\out{\hypertarget{method-CohortTableHandler-getSexFisherTest}{}}}
\subsection{Method \code{getSexFisherTest()}}{
Compares the proportion of males and females in two cohorts using Fisher's exact test.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortTableHandler$getSexFisherTest(
  selected_cohortId1,
  selected_cohortId2,
  testFor = "Subjects"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{selected_cohortId1}}{The cohort id of the first cohort.}

\item{\code{selected_cohortId2}}{The cohort id of the second cohort.}

\item{\code{testFor}}{Character string indicating what to test: "Subjects" or "allEvents". Default is "Subjects".}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a list containing components such as p.value and conf.int of the test

getYearOfBirthTests
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortTableHandler-getYearOfBirthTests"></a>}}
\if{latex}{\out{\hypertarget{method-CohortTableHandler-getYearOfBirthTests}{}}}
\subsection{Method \code{getYearOfBirthTests()}}{
Compares the year of birth distributions between the case and control cohorts using two-sample t-test to compare mean year of birth,
uses cohen's d to assess effect size of the difference in the mean of the year of births,and the Kolmogorov-Smirnov test
to evaluate if year of births in the two cohorts have similar distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortTableHandler$getYearOfBirthTests(
  selected_cohortId1,
  selected_cohortId2,
  testFor = "Subjects"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{selected_cohortId1}}{The cohort id of the first cohort.}

\item{\code{selected_cohortId2}}{The cohort id of the second cohort.}

\item{\code{testFor}}{The type of test to perform. "Subjects" to test the year of birth of the subjects in the cohorts, "allEvents" to test the year of birth of all events in the cohorts.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a list with with three members ttestResult (R htest object), kstestResult (R htest object), cohend result (list of meanInCases, meanInControls, pooledsd, and cohend)

Close Connection
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortTableHandler-closeConnection"></a>}}
\if{latex}{\out{\hypertarget{method-CohortTableHandler-closeConnection}{}}}
\subsection{Method \code{closeConnection()}}{
Closes the database connection and cleans up cohort-specific resources.
This method extends the parent closeConnection to also drop cohort stats tables.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortTableHandler$closeConnection()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortTableHandler-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CohortTableHandler-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortTableHandler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
