% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortsSummary.R
\name{correctEmptyCohortsInCohortsSummary}
\alias{correctEmptyCohortsInCohortsSummary}
\title{Set NA to 0 in a cohortsSummary, and tibles to be empty}
\usage{
correctEmptyCohortsInCohortsSummary(cohortsSummary)
}
\arguments{
\item{cohortsSummary}{tibble in cohortsSummary format}
}
\value{
A tibble with cohort summary structure
}
\description{
Set NA to 0 in a cohortsSummary, and tibles to be empty
}
