% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDMdbHandler.R
\name{createCDMdbHandlerFromList}
\alias{createCDMdbHandlerFromList}
\title{createCDMdbHandlerFromList}
\usage{
createCDMdbHandlerFromList(config)
}
\arguments{
\item{config}{A list containing configuration settings for the CDMdbHandler.
\itemize{
\item databaseName: The name of the database.
\item connection: A list of connection details settings.
\item cdm: A list of CDM database schema settings.
\item cohortTable: The name of the cohort table.
}}
}
\value{
A CDMdbHandler object.
}
\description{
A function to create a CDMdbHandler object from a list of configuration settings.
}
