% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortData.R
\name{createEmptyCohortData}
\alias{createEmptyCohortData}
\title{Create an empty cohort data}
\usage{
createEmptyCohortData()
}
\value{
A tibble with cohort data structure
}
\description{
Create an empty cohort data
}
