% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortsSummary.R
\name{createEmptyCohortsSummary}
\alias{createEmptyCohortsSummary}
\title{Create an empty cohorts summary}
\usage{
createEmptyCohortsSummary()
}
\value{
A tibble with cohort summary structure
}
\description{
Create an empty cohorts summary
}
