% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortGenerator_subsets.R
\name{createOperationSubset}
\alias{createOperationSubset}
\title{A definition of subset functions to be applied to a set of cohorts}
\usage{
createOperationSubset(name = NULL, operationString)
}
\arguments{
\item{name}{optional name of operator}

\item{operationString}{string - the operation string to be used in the query}
}
\value{
a OperationSubsetOperator instance
}
\description{
A definition of subset functions to be applied to a set of cohorts
}
