% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeCodeWAS.R
\name{csvFilesToSqlite}
\alias{csvFilesToSqlite}
\title{csvFilesToSqlite}
\usage{
csvFilesToSqlite(
  dataFolder,
  sqliteDbPath = "timeCodeWAS.sqlite",
  overwrite = FALSE,
  tablePrefix = "",
  analysis = "timeCodeWAS"
)
}
\arguments{
\item{dataFolder}{A character string specifying the folder path where the CSV files are located.}

\item{sqliteDbPath}{A character string specifying the file path for the SQLite database. Default is "timeCodeWAS.sqlite".}

\item{overwrite}{A logical value indicating whether to overwrite the existing SQLite database if it already exists. Default is FALSE.}

\item{tablePrefix}{A character string specifying the prefix to be added to the table names created in the SQLite database.}
}
\value{
A character string indicating the file path of the SQLite database where the data has been imported.
}
\description{
This function reads CSV files from a specified folder and imports them into a SQLite database. It creates tables in the SQLite database based on the specifications provided in a CSV file and uploads the data from the CSV files into these tables.
}
