% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setOperations.R
\name{..binaryTreeToSQL}
\alias{..binaryTreeToSQL}
\title{Convert Binary Tree to SQL Query}
\usage{
..binaryTreeToSQL(binaryTree, side, depth)
}
\arguments{
\item{binaryTree}{A list representing the binary tree structure of an operation.
The list should contain elements 'left', 'operation', and 'right'.}

\item{side}{A string indicating the side of the binary tree node ('left', 'right', 'root').}

\item{depth}{An integer indicating the depth of the current node in the binary tree.}
}
\value{
A list containing the SQL query statement and the position ID of the current node.
}
\description{
Converts a binary tree representation of an operation into SQL query statements.
This function is useful for generating SQL queries from binary tree structures.
}
