% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortData.R
\name{.assertCollectionCohortData}
\alias{.assertCollectionCohortData}
\title{Internal function to validate cohort data}
\usage{
.assertCollectionCohortData(cohortData)
}
\arguments{
\item{cohortData}{The cohort data tibble to be validated}
}
\description{
Internal function to validate cohort data
}
