% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortsSummary.R
\name{.assertCollectionCohortsSummary}
\alias{.assertCollectionCohortsSummary}
\title{Internal function to validate cohort data}
\usage{
.assertCollectionCohortsSummary(CohortsSummary)
}
\arguments{
\item{CohortsSummary}{The cohort data tibble to be validated}
}
\description{
Internal function to validate cohort data
}
