% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable_cohortsSummary.R
\name{.barStyle}
\alias{.barStyle}
\title{Generate Style for Bar}
\usage{
.barStyle(
  perSexStr,
  colorSexMale = "#2c5e77",
  colorSexFemale = "#BF616A",
  colorSexNa = "#8C8C8C"
)
}
\arguments{
\item{perSexStr}{A string representing the percentages of male, female, and NA (e.g., "40\% 30\% 30\%").}

\item{colorSexMale}{The color for the male section (default is "#2c5e77").}

\item{colorSexFemale}{The color for the female section (default is "#BF616A").}

\item{colorSexNa}{The color for the NA section (default is "#8C8C8C").}
}
\value{
A list containing CSS style properties for the bar.
}
\description{
Generates a CSS style for a bar based on the percentage values of male, female, and NA.
}
