% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable_cohortsSummary.R
\name{.buildInfoStr}
\alias{.buildInfoStr}
\title{Converts build information to a string representation.}
\usage{
.buildInfoStr(buildInfo)
}
\arguments{
\item{buildInfo}{Build information containing logs.}
}
\value{
A string representation of build information.
}
\description{
Converts build information to a string representation.
}
