% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable_cohortsSummary.R
\name{.buildInfoToTooltipStr}
\alias{.buildInfoToTooltipStr}
\title{Converts build information to a tooltip string representation.}
\usage{
.buildInfoToTooltipStr(buildInfo)
}
\arguments{
\item{buildInfo}{Build information containing logs.}
}
\value{
A tooltip string representation of build information.
}
\description{
Converts build information to a tooltip string representation.
}
