% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setOperations.R
\name{.operationStringToBinaryTree}
\alias{.operationStringToBinaryTree}
\title{Convert Operation String to Binary Tree Representation}
\usage{
.operationStringToBinaryTree(operationString)
}
\arguments{
\item{operationString}{The input string representing an operation to be converted into a binary tree.}
}
\value{
A list representing the binary tree structure of the operation string.
}
\description{
Converts an operation string into a binary tree representation, where valid operators include "Upd", "Ip", and "Mp".
This function is useful for parsing operation strings and constructing binary trees for evaluation.
}
