% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable_cohortsSummary.R
\name{.sexTibbleToStr}
\alias{.sexTibbleToStr}
\title{Convert Sex Data to String}
\usage{
.sexTibbleToStr(data)
}
\arguments{
\item{data}{A tibble containing sex data.}
}
\value{
A string representation of the sex data (e.g., "40\% 30\% 30\%").
}
\description{
Converts sex data in a tibble to a string representation.
}
