% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable_cohortsSummary.R
\name{.sexTibbleToTooltipStr}
\alias{.sexTibbleToTooltipStr}
\title{Convert Sex Data to Tooltip String}
\usage{
.sexTibbleToTooltipStr(data)
}
\arguments{
\item{data}{A tibble containing sex data.}
}
\value{
A tooltip string representation of the sex data.
}
\description{
Converts sex data in a tibble to a tooltip string representation.
}
