% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortTableHandler.R
\name{.solveShortNameConflicts}
\alias{.solveShortNameConflicts}
\title{Resolve Cohort Short Name Conflicts}
\usage{
.solveShortNameConflicts(newShortNames, existingShortNames)
}
\arguments{
\item{newShortNames}{Character vector. The newly generated short names to be made unique.}

\item{existingShortNames}{Character vector. The set of existing short names to avoid conflicts with.}
}
\value{
A character vector of unique short names.
}
\description{
Given a set of new and existing short names, resolve any duplicates. Appends an index to any duplicate names to ensure uniqueness.
}
