% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable_cohortsSummary.R
\name{.tippyText}
\alias{.tippyText}
\title{Generate Tippy Text}
\usage{
.tippyText(text, tooltip)
}
\arguments{
\item{text}{The text to display.}

\item{tooltip}{The tooltip text.}
}
\value{
A tippy object.
}
\description{
Generates tippy text with a tooltip for a given text.
}
