% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortDemographics.R
\name{executeCohortDemographicsCounts}
\alias{executeCohortDemographicsCounts}
\title{executedemographicsCounts}
\usage{
executeCohortDemographicsCounts(
  exportFolder,
  cohortTableHandler = NULL,
  cohortIds,
  referenceYears = c("cohort_start_date", "cohort_end_date", "birth_datetime"),
  groupBy = c("calendarYear", "ageGroup", "gender"),
  minCellCount = 0
)
}
\arguments{
\item{exportFolder}{A character string specifying the folder path where the results will be exported.}

\item{cohortTableHandler}{An object of class 'CohortTableHandler' representing the cohort table handler.}

\item{minCellCount}{An integer specifying the minimum cell count for Fisher's exact test.}

\item{cohortIdCases}{The ID of the cohort representing cases.}

\item{cohortIdControls}{The ID of the cohort representing controls.}

\item{covariateSettings}{A list of settings for extracting temporal covariates.}
}
\value{
A logical value indicating if the function was executed successfully.
}
\description{
This function calculates the number of subjects with observation case and controls in each time window for a given cohort. It returns a data frame with the counts of cases and controls for each time window and covariate, as well as the results of a Fisher's exact test comparing the counts of cases and controls.
}
