% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureExtraction_covariates.R
\name{FeatureExtraction_createTemporalCovariateSettingsFromList}
\alias{FeatureExtraction_createTemporalCovariateSettingsFromList}
\title{Create Temporal Covariate Settings From List}
\usage{
FeatureExtraction_createTemporalCovariateSettingsFromList(
  analysisIds,
  temporalStartDays = c(-99999),
  temporalEndDays = c(99999)
)
}
\arguments{
\item{analysisIds}{A vector of analysis IDs.}

\item{temporalStartDays}{A vector of temporal start days for each analysis ID.}

\item{temporalEndDays}{A vector of temporal end days for each analysis ID.}
}
\value{
A list of temporal covariate settings.
}
\description{
This function creates temporal covariate settings based on a list of analysis IDs, temporal start days, and temporal end days.
}
