% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureExtraction_analysisDetails.R
\name{FeatureExtraction_createTemporalSourceCovariateSettings}
\alias{FeatureExtraction_createTemporalSourceCovariateSettings}
\title{Create Temporal Source Covariate Settings}
\usage{
FeatureExtraction_createTemporalSourceCovariateSettings(
  useConditionOccurrenceSourceConcept = TRUE,
  useDrugExposureSourceConcept = TRUE,
  useProcedureOccurrenceSourceConcept = TRUE,
  useMeasurementSourceConcept = TRUE,
  useDeviceExposureSourceConcept = TRUE,
  useObservationSourceConcept = TRUE,
  temporalStartDays = -365:-1,
  temporalEndDays = -365:-1
)
}
\arguments{
\item{useConditionOccurrenceSourceConcept}{Logical indicating whether to include condition occurrence source concepts.}

\item{useDrugExposureSourceConcept}{Logical indicating whether to include drug exposure source concepts.}

\item{useProcedureOccurrenceSourceConcept}{Logical indicating whether to include procedure occurrence source concepts.}

\item{useMeasurementSourceConcept}{Logical indicating whether to include measurement source concepts.}

\item{useDeviceExposureSourceConcept}{Logical indicating whether to include device exposure source concepts.}

\item{useObservationSourceConcept}{Logical indicating whether to include observation source concepts.}

\item{temporalStartDays}{Vector of integers representing the start days of the temporal window for covariate extraction.}

\item{temporalEndDays}{Vector of integers representing the end days of the temporal window for covariate extraction.}
}
\value{
Settings for temporal source covariates suitable for feature extraction.
}
\description{
This function generates settings for temporal source covariates to be used in feature extraction, DEPRECATED use FeatureExtraction_createTemporalCovariateSettingsFromList.
}
\details{
This function generates settings for temporal source covariates, which can be used in subsequent feature extraction tasks. It allows specifying which types of source concepts to include and the temporal window for extraction.
}
