% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortData.R
\name{getCohortDataFromCohortTable}
\alias{getCohortDataFromCohortTable}
\title{Get Cohort Data from Cohort Table}
\usage{
getCohortDataFromCohortTable(
  connectionDetails = NULL,
  connection = NULL,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable,
  cohortNameIds
)
}
\arguments{
\item{connectionDetails}{A list of connection details (optional).}

\item{connection}{An existing database connection (optional).}

\item{cdmDatabaseSchema}{The schema name for the CDM database.}

\item{cohortDatabaseSchema}{The schema name for the cohort database.}

\item{cohortTable}{The name of the cohort table.}

\item{cohortNameIds}{A data frame containing cohort name IDs.}
}
\value{
Returns TRUE if the cohort data is successfully retrieved.
}
\description{
This function retrieves cohort data from a cohort table in a database.
}
\details{
This function retrieves cohort data from a specified cohort table in a database.
It validates the input parameters, establishes a database connection if one is not provided,
and then performs SQL operations to retrieve the cohort data.
}
