% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortData.R
\name{getCohortDataReadrColTypes}
\alias{getCohortDataReadrColTypes}
\title{getCohortDataReadrColTypes}
\usage{
getCohortDataReadrColTypes()
}
\value{
A readr column types object.
}
\description{
Returns a readr column types object for the cohort data.
}
