% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortDemographics.R
\name{getCohortDemographicsCounts}
\alias{getCohortDemographicsCounts}
\title{Retrieve Cohort Demographics Counts}
\usage{
getCohortDemographicsCounts(
  connectionDetails = NULL,
  connection = NULL,
  cdmDatabaseSchema,
  vocabularyDatabaseSchema = cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable = "cohort",
  cohortIds = c(),
  referenceYears = c("cohort_start_date", "cohort_end_date", "birth_datetime")
)
}
\arguments{
\item{connectionDetails}{An object of class \code{DatabaseConnector::connectionDetails}. This provides the details required to connect to the database. Either \code{connectionDetails} or \code{connection} must be provided.}

\item{connection}{A database connection object. If not provided, the function will establish a connection using \code{connectionDetails}.}

\item{cdmDatabaseSchema}{The schema name where the Common Data Model (CDM) data resides. This is used to locate the necessary tables in the database.}

\item{vocabularyDatabaseSchema}{The schema name where the vocabulary tables reside. Defaults to the value of \code{cdmDatabaseSchema} if not specified.}

\item{cohortDatabaseSchema}{The schema name where the cohort tables reside. This schema contains the cohort data to be analyzed.}

\item{cohortTable}{The name of the cohort table. Defaults to \code{"cohort"}.}

\item{cohortIds}{A numeric vector of cohort IDs for which the demographic counts will be retrieved. If empty, no cohorts will be analyzed.}

\item{referenceYears}{A character vector specifying the reference year columns to be used for the demographic counts. Must be one or more of \code{"cohort_start_date"}, \code{"cohort_end_date"}, or \code{"birth_datetime"}.}
}
\value{
A tibble containing the aggregated demographic counts, including columns for \code{referenceYear}, \code{cohortId}, \code{calendarYear}, \code{ageGroup}, \code{gender}, and \code{count}.
}
\description{
This function retrieves and aggregates demographic counts for specified cohorts based on reference years. It connects to the database, executes SQL queries to calculate demographic statistics such as age group and gender distribution, and returns the results as a tibble.
}
\details{
The function performs the following steps:
\enumerate{
\item Validates the input parameters.
\item Establishes a database connection if one is not already provided.
\item Iterates over the \code{referenceYears} to calculate demographic counts for each cohort.
\item Translates the SQL script according to the target database dialect and executes it.
\item Aggregates the results by age group, gender, and calendar year.
}
}
