% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortTable.R
\name{getCohortNamesFromCohortDefinitionTable}
\alias{getCohortNamesFromCohortDefinitionTable}
\title{Get Cohort Names from Cohort Definition Table}
\usage{
getCohortNamesFromCohortDefinitionTable(
  connectionDetails = NULL,
  connection = NULL,
  cohortDatabaseSchema,
  cohortDefinitionTable = "cohort_definition"
)
}
\arguments{
\item{connectionDetails}{Details required to establish a database connection (optional).}

\item{connection}{An existing database connection (optional).}

\item{cohortDatabaseSchema}{The database schema where the cohort definition table is expected to be found.}

\item{cohortDefinitionTable}{The name of the cohort definition table. Defaults to "cohort_definition".}
}
\value{
A tibble containing cohort names, IDs, and descriptions.
}
\description{
Retrieves cohort names along with their corresponding IDs and descriptions from the specified cohort definition table
within a given database schema. This function assists in querying cohort information from the database.
}
