% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureExtraction_covariates.R
\name{getListOfAnalysis}
\alias{getListOfAnalysis}
\title{Get List of Analysis}
\usage{
getListOfAnalysis()
}
\value{
A data frame containing analysis information.
}
\description{
This function returns a tibble containing analysis information.
}
