% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HadesExtras-package.R
\docType{package}
\name{HadesExtras-package}
\alias{HadesExtras}
\alias{HadesExtras-package}
\title{HadesExtras: HadesExtras}
\description{
Provides additional functions to complement Hades packages for observational health data analysis
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/FINNGEN/HadesExtras}
  \item \url{https://finngen.github.io/HadesExtras/}
  \item Report bugs at \url{https://github.com/FINNGEN/HadesExtras/issues}
}

}
\author{
\strong{Maintainer}: Javier Gracia-Tabuenca \email{javier.graciatabuenca@tuni.fi} (\href{https://orcid.org/0000-0002-2455-0598}{ORCID})

}
\keyword{internal}
