% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortTableHandler.R
\name{makeShortName}
\alias{makeShortName}
\title{Cohort Short Name Utilities}
\usage{
makeShortName(name, id)
}
\arguments{
\item{name}{Character. The full name of the cohort.}

\item{id}{Integer or character. The id of the cohort.}
}
\value{
\code{makeShortName}: A character value representing the short cohort name.
}
\description{
Function to generate concise, human-readable short names for cohorts,
and to resolve conflicts to ensure uniqueness of cohort short names.
}
