% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setOperations.R
\name{operationStringToSQL}
\alias{operationStringToSQL}
\title{Convert Operation String to SQL Query}
\usage{
operationStringToSQL(operationString)
}
\arguments{
\item{operationString}{The input string representing an operation to be converted into SQL.}
}
\value{
An SQL query string derived from the input operation string.
}
\description{
Converts an operation string into an SQL query that can be executed against a database.
This function first converts the operation string into a binary tree representation and then
generates the corresponding SQL query.
}
