% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfigurationHelpers.R
\name{readAndParseYalm}
\alias{readAndParseYalm}
\title{Read and Parse YAML File with Placeholder Replacement}
\usage{
readAndParseYalm(pathToYalmFile, ...)
}
\arguments{
\item{pathToYalmFile}{A string representing the file path to the YAML file.}

\item{...}{Named arguments to replace placeholders in the format \verb{<name>} within the YAML file.}
}
\value{
A parsed list representing the contents of the modified YAML file.
}
\description{
Reads a YAML file from a given path, replaces specified placeholders with provided values,
and returns the parsed content. If any provided placeholders are not found in the YAML file,
the function throws an error.
}
\examples{
# Assuming the YAML file contains placeholders like <name>:
# name: <name>
# age: <age>
#
# You can replace those placeholders by passing named arguments:
readAndParseYalm("path/to/yaml/file", name = "John", age = 30)

}
