% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortData.R
\name{readCohortData}
\alias{readCohortData}
\title{readCohortData}
\usage{
readCohortData(pathCohortDataFile, delim = ",")
}
\arguments{
\item{pathCohortDataFile}{The path to the cohort data file.}

\item{delim}{The delimiter used in the cohort data file.}
}
\value{
A data frame containing the cohort data.
}
\description{
Reads a cohort data file and returns a data frame.
}
