% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_DatabasesSettings.R
\name{readDatabaseSettings}
\alias{readDatabaseSettings}
\title{Title readDatabaseSettings}
\usage{
readDatabaseSettings(path_databases_settings_yalm, database_name)
}
\arguments{
\item{path_databases_settings_yalm}{path to the databases settings yalm file}

\item{database_name}{a valid database name in the databases settings yalm file}
}
\value{
database settings
}
\description{
Returns the database settings for the \code{database_name} in the databases settings yalm file.
Rewrites the connectionDetails using  DatabaseConnector::createConnectionDetails
}
