% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_DatabasesSettings.R
\name{readDatabasesNames}
\alias{readDatabasesNames}
\title{readDatabasesNames}
\usage{
readDatabasesNames(path_databases_settings_yalm)
}
\arguments{
\item{path_databases_settings_yalm}{path to the databases settings yalm file}
}
\value{
database names
}
\description{
Return database names available in a databases settings yalm file
}
