% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable_cohortsSummary.R
\name{rectable_cohortsSummary}
\alias{rectable_cohortsSummary}
\title{Cohorts Summary Table}
\usage{
rectable_cohortsSummary(cohortsSummary, deleteButtonsShinyId = NULL)
}
\arguments{
\item{cohortsSummary}{A tibble in  cohortsSummary format.}

\item{deleteButtonsShinyId}{An optional Shiny input ID for handling the click of the delete buttons.}
}
\value{
A reactable table displaying cohort summary information.
}
\description{
Generates a summary table of cohorts, including various cohort statistics.
}
