% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setOperations.R
\name{splitString}
\alias{splitString}
\title{Split String into Valid Tokens}
\usage{
splitString(string)
}
\arguments{
\item{string}{The input string to be split into tokens.}
}
\value{
A character vector containing valid tokens extracted from the input string.
}
\description{
Splits a string into valid tokens based on certain rules and operators. Valid tokens include "Upd", "Ip", "Mp",
numeric values, and parentheses. This function is useful for parsing strings containing a cohort operation.
}
