% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp_fixes.R
\name{dbplyr::in_schema}
\alias{dbplyr::in_schema}
\title{tmp fix for DatabaseConnector::inDatabaseSchema}
\usage{
dbplyr::in_schema(databaseSchema, table)
}
\arguments{
\item{databaseSchema}{The name of the database schema.}

\item{table}{The name of the table.}
}
\value{
The fully qualified table name with the database schema.
}
\description{
till this is fixed https://github.com/OHDSI/DatabaseConnector/issues/236
}
