% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDMdbHandler.R
\name{CDMdbHandler}
\alias{CDMdbHandler}
\title{CDMdbHandler}
\description{
Class for handling database connection and schema information for a CDM database
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{connectionHandler}}{ConnectionHandler object for managing the database connection}

\item{\code{vocabularyDatabaseSchema}}{Name of the vocabulary database schema}

\item{\code{cdmDatabaseSchema}}{Name of the CDM database schema}

\item{\code{connectionStatusLog}}{Log tibble object for storing connection status information}

\item{\code{vocabularyInfo}}{Data frame containing information about the vocabulary database}

\item{\code{CDMInfo}}{Data frame containing information about the CDM database}

\item{\code{getTblVocabularySchema}}{List of functions that create dbplyr table for the vocabulary tables}

\item{\code{getTblCDMSchema}}{List of functions that create dbplyr table for the CDM tables}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CDMdbHandler-new}{\code{CDMdbHandler$new()}}
\item \href{#method-CDMdbHandler-finalize}{\code{CDMdbHandler$finalize()}}
\item \href{#method-CDMdbHandler-loadConnection}{\code{CDMdbHandler$loadConnection()}}
\item \href{#method-CDMdbHandler-clone}{\code{CDMdbHandler$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CDMdbHandler-new"></a>}}
\if{latex}{\out{\hypertarget{method-CDMdbHandler-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CDMdbHandler$new(
  connectionHandler,
  cdmDatabaseSchema,
  vocabularyDatabaseSchema = cdmDatabaseSchema
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{connectionHandler}}{A ConnectionHandler object}

\item{\code{cdmDatabaseSchema}}{Name of the CDM database schema}

\item{\code{vocabularyDatabaseSchema}}{(Optional) Name of the vocabulary database schema (default is cdmDatabaseSchema)
Finalize method}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CDMdbHandler-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-CDMdbHandler-finalize}{}}}
\subsection{Method \code{finalize()}}{
Closes the connection if active.
Reload connection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CDMdbHandler$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CDMdbHandler-loadConnection"></a>}}
\if{latex}{\out{\hypertarget{method-CDMdbHandler-loadConnection}{}}}
\subsection{Method \code{loadConnection()}}{
Updates the connection status by checking the database connection, vocabulary database schema, and CDM database schema.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CDMdbHandler$loadConnection()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CDMdbHandler-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CDMdbHandler-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CDMdbHandler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
