% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortData.R
\name{cohortDataToCohortDefinitionSet}
\alias{cohortDataToCohortDefinitionSet}
\title{cohortDataToCohortDefinitionSet}
\usage{
cohortDataToCohortDefinitionSet(
  connectionDetails = NULL,
  connection = NULL,
  cdmDatabaseSchema,
  cohortData
)
}
\arguments{
\item{connectionDetails}{The connection details to the database. If not connection is provided.}

\item{connection}{The database connection. If not connectionDetails is provided.}

\item{cdmDatabaseSchema}{The schema of the CDM database.}

\item{cohortData}{The cohort data.}
}
\value{
The cohortDefinitionSet with additional column
}
\description{
Generates a cohortDefinitionSet based on cohortData tibble.
Finds the person_id based on the person_source_value in cohortData.
If cohort_start_date and cohort_end_date is missing in cohortData,
it sets the first observation_period_start_date and last observation_period_end_date respectvively. #'
The cohortDefinitionSet includes a sql to copy from a temporary table with the pre-processed cohortData and
a tibble with information on the pre-processing.
}
