% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortData.R
\name{cohortDataToCohortDefinitionSet}
\alias{cohortDataToCohortDefinitionSet}
\title{cohortDataToCohortDefinitionSet}
\usage{
cohortDataToCohortDefinitionSet(
  cohortData,
  cohortIdOffset = 0L,
  skipCohortDataCheck = FALSE
)
}
\arguments{
\item{cohortData}{The tibble in cohortData format.}

\item{cohortIdOffset}{An integer offset for cohort IDs (default is 0L).}

\item{skipCohortDataCheck}{Logical value indicating whether to skip cohort data validation (default is FALSE).}
}
\value{
A data frame of format cohortDefinitionSet.
}
\description{
Convert a cohortData tibble into a cohortDefinitionSet
}
