% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortGenerator.R
\name{CohortGenerator_createCohortTables}
\alias{CohortGenerator_createCohortTables}
\title{CohortGenerator_createCohortTables}
\usage{
CohortGenerator_createCohortTables(
  connectionDetails = NULL,
  connection = NULL,
  cohortDatabaseSchema,
  cohortTableNames = CohortGenerator::getCohortTableNames(),
  incremental = FALSE
)
}
\description{
Wrapper for CohortGenerator::createCohortTables, where if bigquery is used, the tables are created using bigrquery::bq_table_create.
}
