% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortGenerator.R
\name{CohortGenerator_deleteCohortFromCohortTable}
\alias{CohortGenerator_deleteCohortFromCohortTable}
\title{deleteCohortFromCohortTable}
\usage{
CohortGenerator_deleteCohortFromCohortTable(
  connectionDetails = NULL,
  connection = NULL,
  cohortDatabaseSchema,
  cohortTableNames,
  cohortIds
)
}
\arguments{
\item{connectionDetails}{A list of database connection details (optional).}

\item{connection}{A pre-established database connection (optional).}

\item{cohortDatabaseSchema}{The schema name of the cohort database.}

\item{cohortTableNames}{A list containing the name of the cohort table.}

\item{cohortIds}{Numeric vector of cohort IDs to be deleted.}
}
\value{
TRUE if the deletion was successful; otherwise, an error is raised.
}
\description{
Deletes specified cohorts from a cohort table in a database.
}
