% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortGenerator.R
\name{CohortGenerator_generateCohortSet}
\alias{CohortGenerator_generateCohortSet}
\title{Wrap around CohortGenerator::generateCohortSet}
\usage{
CohortGenerator_generateCohortSet(
  connectionDetails = NULL,
  connection = NULL,
  cdmDatabaseSchema,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  cohortDatabaseSchema = cdmDatabaseSchema,
  cohortTableNames = getCohortTableNames(),
  cohortDefinitionSet = NULL,
  stopOnError = TRUE,
  incremental = FALSE,
  incrementalFolder = NULL
)
}
\arguments{
\item{connectionDetails}{as in CohortGenerator::generateCohortSet.}

\item{connection}{as in CohortGenerator::generateCohortSet.}

\item{cdmDatabaseSchema}{as in CohortGenerator::generateCohortSet.}

\item{tempEmulationSchema}{as in CohortGenerator::generateCohortSet.}

\item{cohortDatabaseSchema}{as in CohortGenerator::generateCohortSet.}

\item{cohortTableNames}{as in CohortGenerator::generateCohortSet.}

\item{cohortDefinitionSet}{as in CohortGenerator::generateCohortSet.}

\item{stopOnError}{as in CohortGenerator::generateCohortSet.}

\item{incremental}{as in CohortGenerator::generateCohortSet.}

\item{incrementalFolder}{as in CohortGenerator::generateCohortSet.}
}
\value{
resutls from CohortGenerator::generateCohortSet with additional column 'buildInfo'
}
\description{
Wrap around CohortGenerator::generateCohortSet to be able to build cohorts based on cohortData.
Checks if cohortDefinitionSet contais a cohort based on cohortData.
If so, uploads cohortData found in the json column to the database and transforms that data in to a temporal table with cohortTable format.
CohortGenerator::generateCohortSet is called and processes the sql column that will copy the contents of the temporal cohortTable to the
target cohortTable.
It adds a column to the restuls tibble, 'buildInfo' with a LogTibble with information on the cohort construction.
}
