% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortGenerator_subsets.R
\name{CohortGenerator_MatchingSubsetOperator}
\alias{CohortGenerator_MatchingSubsetOperator}
\title{Matching Subset Operator}
\description{
A subset of type cohort - subset a population to only those contained within defined cohort
to List
}
\section{Super class}{
\code{\link[CohortGenerator:SubsetOperator]{CohortGenerator::SubsetOperator}} -> \code{MatchingSubsetOperator}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{matchToCohortId}}{Integer id of cohorts to match to}

\item{\code{matchRatio}}{matching ratio}

\item{\code{matchSex}}{match to target sex}

\item{\code{matchBirthYear}}{match to target birth year}

\item{\code{matchCohortStartDateWithInDuration}}{match have start date as in cohort duration}

\item{\code{newCohortStartDate}}{change cohort start date to the matched person}

\item{\code{newCohortEndDate}}{change cohort end date to the matched person}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MatchingSubsetOperator-toList}{\code{CohortGenerator_MatchingSubsetOperator$toList()}}
\item \href{#method-MatchingSubsetOperator-getAutoGeneratedName}{\code{CohortGenerator_MatchingSubsetOperator$getAutoGeneratedName()}}
\item \href{#method-MatchingSubsetOperator-clone}{\code{CohortGenerator_MatchingSubsetOperator$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="CohortGenerator" data-topic="SubsetOperator" data-id="classname"><a href='../../CohortGenerator/html/SubsetOperator.html#method-SubsetOperator-classname'><code>CohortGenerator::SubsetOperator$classname()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CohortGenerator" data-topic="SubsetOperator" data-id="getQueryBuilder"><a href='../../CohortGenerator/html/SubsetOperator.html#method-SubsetOperator-getQueryBuilder'><code>CohortGenerator::SubsetOperator$getQueryBuilder()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CohortGenerator" data-topic="SubsetOperator" data-id="initialize"><a href='../../CohortGenerator/html/SubsetOperator.html#method-SubsetOperator-initialize'><code>CohortGenerator::SubsetOperator$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CohortGenerator" data-topic="SubsetOperator" data-id="isEqualTo"><a href='../../CohortGenerator/html/SubsetOperator.html#method-SubsetOperator-isEqualTo'><code>CohortGenerator::SubsetOperator$isEqualTo()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CohortGenerator" data-topic="SubsetOperator" data-id="publicFields"><a href='../../CohortGenerator/html/SubsetOperator.html#method-SubsetOperator-publicFields'><code>CohortGenerator::SubsetOperator$publicFields()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CohortGenerator" data-topic="SubsetOperator" data-id="toJSON"><a href='../../CohortGenerator/html/SubsetOperator.html#method-SubsetOperator-toJSON'><code>CohortGenerator::SubsetOperator$toJSON()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MatchingSubsetOperator-toList"></a>}}
\if{latex}{\out{\hypertarget{method-MatchingSubsetOperator-toList}{}}}
\subsection{Method \code{toList()}}{
List representation of object
Get auto generated name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortGenerator_MatchingSubsetOperator$toList()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MatchingSubsetOperator-getAutoGeneratedName"></a>}}
\if{latex}{\out{\hypertarget{method-MatchingSubsetOperator-getAutoGeneratedName}{}}}
\subsection{Method \code{getAutoGeneratedName()}}{
name generated from subset operation properties
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortGenerator_MatchingSubsetOperator$getAutoGeneratedName()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
character
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MatchingSubsetOperator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MatchingSubsetOperator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortGenerator_MatchingSubsetOperator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
