% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortGenerator_subsets.R
\name{CohortGenerator_OperationSubsetOperator}
\alias{CohortGenerator_OperationSubsetOperator}
\title{Operation Subset Operator}
\description{
A subset of type cohort - subset a population to only those contained within defined cohort
to List
}
\section{Super class}{
\code{\link[CohortGenerator:SubsetOperator]{CohortGenerator::SubsetOperator}} -> \code{OperationSubsetOperator}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{operationString}}{Integer id of cohorts to match to}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OperationSubsetOperator-toList}{\code{CohortGenerator_OperationSubsetOperator$toList()}}
\item \href{#method-OperationSubsetOperator-getAutoGeneratedName}{\code{CohortGenerator_OperationSubsetOperator$getAutoGeneratedName()}}
\item \href{#method-OperationSubsetOperator-clone}{\code{CohortGenerator_OperationSubsetOperator$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="CohortGenerator" data-topic="SubsetOperator" data-id="classname"><a href='../../CohortGenerator/html/SubsetOperator.html#method-SubsetOperator-classname'><code>CohortGenerator::SubsetOperator$classname()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CohortGenerator" data-topic="SubsetOperator" data-id="getQueryBuilder"><a href='../../CohortGenerator/html/SubsetOperator.html#method-SubsetOperator-getQueryBuilder'><code>CohortGenerator::SubsetOperator$getQueryBuilder()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CohortGenerator" data-topic="SubsetOperator" data-id="initialize"><a href='../../CohortGenerator/html/SubsetOperator.html#method-SubsetOperator-initialize'><code>CohortGenerator::SubsetOperator$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CohortGenerator" data-topic="SubsetOperator" data-id="isEqualTo"><a href='../../CohortGenerator/html/SubsetOperator.html#method-SubsetOperator-isEqualTo'><code>CohortGenerator::SubsetOperator$isEqualTo()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CohortGenerator" data-topic="SubsetOperator" data-id="publicFields"><a href='../../CohortGenerator/html/SubsetOperator.html#method-SubsetOperator-publicFields'><code>CohortGenerator::SubsetOperator$publicFields()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CohortGenerator" data-topic="SubsetOperator" data-id="toJSON"><a href='../../CohortGenerator/html/SubsetOperator.html#method-SubsetOperator-toJSON'><code>CohortGenerator::SubsetOperator$toJSON()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OperationSubsetOperator-toList"></a>}}
\if{latex}{\out{\hypertarget{method-OperationSubsetOperator-toList}{}}}
\subsection{Method \code{toList()}}{
List representation of object
Get auto generated name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortGenerator_OperationSubsetOperator$toList()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OperationSubsetOperator-getAutoGeneratedName"></a>}}
\if{latex}{\out{\hypertarget{method-OperationSubsetOperator-getAutoGeneratedName}{}}}
\subsection{Method \code{getAutoGeneratedName()}}{
name generated from subset operation properties
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortGenerator_OperationSubsetOperator$getAutoGeneratedName()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
character
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OperationSubsetOperator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OperationSubsetOperator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortGenerator_OperationSubsetOperator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
