% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortTableHandler.R
\name{CohortTableHandler}
\alias{CohortTableHandler}
\title{CohortTableHandler}
\description{
Class for handling cohort tables in a CDM database.
Inherits from CDMdbHandler.
}
\section{Super class}{
\code{\link[HadesExtras:CDMdbHandler]{HadesExtras::CDMdbHandler}} -> \code{CohortTableHandler}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{cohortDatabaseSchema}}{Name of the cohort database schema.}

\item{\code{cohortTableNames}}{Names of the cohort tables.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CohortTableHandler-new}{\code{CohortTableHandler$new()}}
\item \href{#method-CohortTableHandler-finalize}{\code{CohortTableHandler$finalize()}}
\item \href{#method-CohortTableHandler-loadConnection}{\code{CohortTableHandler$loadConnection()}}
\item \href{#method-CohortTableHandler-insertOrUpdateCohorts}{\code{CohortTableHandler$insertOrUpdateCohorts()}}
\item \href{#method-CohortTableHandler-deleteCohorts}{\code{CohortTableHandler$deleteCohorts()}}
\item \href{#method-CohortTableHandler-getCohortCounts}{\code{CohortTableHandler$getCohortCounts()}}
\item \href{#method-CohortTableHandler-getCohortsSummary}{\code{CohortTableHandler$getCohortsSummary()}}
\item \href{#method-CohortTableHandler-getCohortNames}{\code{CohortTableHandler$getCohortNames()}}
\item \href{#method-CohortTableHandler-clone}{\code{CohortTableHandler$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortTableHandler-new"></a>}}
\if{latex}{\out{\hypertarget{method-CohortTableHandler-new}{}}}
\subsection{Method \code{new()}}{
Initialize the CohortTableHandler object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortTableHandler$new(
  connectionHandler,
  cdmDatabaseSchema,
  vocabularyDatabaseSchema = cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTableName
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{connectionHandler}}{The connection handler object.}

\item{\code{cdmDatabaseSchema}}{Name of the CDM database schema.}

\item{\code{vocabularyDatabaseSchema}}{Name of the vocabulary database schema. Default is the same as the CDM database schema.}

\item{\code{cohortDatabaseSchema}}{Name of the cohort database schema.}

\item{\code{cohortTableName}}{Name of the cohort table.
Finalize method}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortTableHandler-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-CohortTableHandler-finalize}{}}}
\subsection{Method \code{finalize()}}{
Closes the connection if active.

loadConnection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortTableHandler$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortTableHandler-loadConnection"></a>}}
\if{latex}{\out{\hypertarget{method-CohortTableHandler-loadConnection}{}}}
\subsection{Method \code{loadConnection()}}{
Reloads the connection with the initial setting and updates connection status

insertOrUpdateCohorts
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortTableHandler$loadConnection()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortTableHandler-insertOrUpdateCohorts"></a>}}
\if{latex}{\out{\hypertarget{method-CohortTableHandler-insertOrUpdateCohorts}{}}}
\subsection{Method \code{insertOrUpdateCohorts()}}{
If there is no cohort with the same name it is added and generated with a new cohort_id, given cohort_id is saved as source_cohort_id,
If there is a cohort with the same name, the current cohort_id is kept and source_cohort_id is updated
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortTableHandler$insertOrUpdateCohorts(cohortDefinitionSet)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cohortDefinitionSet}}{The cohort definition set to add.

deleteCohorts}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortTableHandler-deleteCohorts"></a>}}
\if{latex}{\out{\hypertarget{method-CohortTableHandler-deleteCohorts}{}}}
\subsection{Method \code{deleteCohorts()}}{
Deletes cohorts from the cohort table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortTableHandler$deleteCohorts(cohortIds)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cohortIds}}{The cohort ids to delete.

getCohortCounts}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortTableHandler-getCohortCounts"></a>}}
\if{latex}{\out{\hypertarget{method-CohortTableHandler-getCohortCounts}{}}}
\subsection{Method \code{getCohortCounts()}}{
Retrieves cohort counts from the cohort table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortTableHandler$getCohortCounts()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A tibble containing the cohort counts with names.

getCohortsSummary
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortTableHandler-getCohortsSummary"></a>}}
\if{latex}{\out{\hypertarget{method-CohortTableHandler-getCohortsSummary}{}}}
\subsection{Method \code{getCohortsSummary()}}{
Retrieves the summary of cohorts including cohort start and end year histograms and sex counts.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortTableHandler$getCohortsSummary()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A tibble containing cohort summary.

getCohortNames
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortTableHandler-getCohortNames"></a>}}
\if{latex}{\out{\hypertarget{method-CohortTableHandler-getCohortNames}{}}}
\subsection{Method \code{getCohortNames()}}{
Retrieves the cohort names.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortTableHandler$getCohortNames()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A vector with the name of the cohorts
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortTableHandler-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CohortTableHandler-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortTableHandler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
