% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortTableHandler.R
\name{CohortTableHandler}
\alias{CohortTableHandler}
\title{CohortTableHandler}
\description{
Class for handling cohort tables in a CDM database.
Inherits from CDMdbHandler.
}
\section{Super class}{
\code{\link[HadesExtras:CDMdbHandler]{HadesExtras::CDMdbHandler}} -> \code{CohortTableHandler}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{cohortDatabaseSchema}}{Name of the cohort database schema (read-only).}

\item{\code{cohortTableNames}}{Names of the cohort tables (read-only).}

\item{\code{incrementalFolder}}{Path to folder used by CohortGenerator in inclemetnal mode (read-only).}

\item{\code{cohortDefinitionSet}}{Table in cohortDefinitionSet with the current cohorts in the cohortTable (read-only).}

\item{\code{cohortGeneratorResults}}{Table with results from CohortGenerator_generateCohortSet with the current cohorts in the cohortTable(read-only).}

\item{\code{cohortDemograpics}}{Table with results from CohortGenerator_getCohortDemograpics with the current cohorts in the cohortTable(read-only).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CohortTableHandler-new}{\code{CohortTableHandler$new()}}
\item \href{#method-CohortTableHandler-finalize}{\code{CohortTableHandler$finalize()}}
\item \href{#method-CohortTableHandler-loadConnection}{\code{CohortTableHandler$loadConnection()}}
\item \href{#method-CohortTableHandler-insertOrUpdateCohorts}{\code{CohortTableHandler$insertOrUpdateCohorts()}}
\item \href{#method-CohortTableHandler-deleteCohorts}{\code{CohortTableHandler$deleteCohorts()}}
\item \href{#method-CohortTableHandler-getCohortCounts}{\code{CohortTableHandler$getCohortCounts()}}
\item \href{#method-CohortTableHandler-getCohortsSummary}{\code{CohortTableHandler$getCohortsSummary()}}
\item \href{#method-CohortTableHandler-getCohortIdAndNames}{\code{CohortTableHandler$getCohortIdAndNames()}}
\item \href{#method-CohortTableHandler-clone}{\code{CohortTableHandler$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="HadesExtras" data-topic="CDMdbHandler" data-id="getConnectionStatus"><a href='../../HadesExtras/html/CDMdbHandler.html#method-CDMdbHandler-getConnectionStatus'><code>HadesExtras::CDMdbHandler$getConnectionStatus()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortTableHandler-new"></a>}}
\if{latex}{\out{\hypertarget{method-CohortTableHandler-new}{}}}
\subsection{Method \code{new()}}{
Initialize the CohortTableHandler object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortTableHandler$new(
  connectionHandler,
  databaseName,
  cdmDatabaseSchema,
  vocabularyDatabaseSchema = cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTableName
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{connectionHandler}}{The connection handler object.}

\item{\code{databaseName}}{A text id for the database the it connects to.}

\item{\code{cdmDatabaseSchema}}{Name of the CDM database schema.}

\item{\code{vocabularyDatabaseSchema}}{Name of the vocabulary database schema. Default is the same as the CDM database schema.}

\item{\code{cohortDatabaseSchema}}{Name of the cohort database schema.}

\item{\code{cohortTableName}}{Name of the cohort table.
Finalize method}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortTableHandler-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-CohortTableHandler-finalize}{}}}
\subsection{Method \code{finalize()}}{
Closes the connection if active.

loadConnection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortTableHandler$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortTableHandler-loadConnection"></a>}}
\if{latex}{\out{\hypertarget{method-CohortTableHandler-loadConnection}{}}}
\subsection{Method \code{loadConnection()}}{
Reloads the connection with the initial setting and updates connection status

insertOrUpdateCohorts
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortTableHandler$loadConnection()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortTableHandler-insertOrUpdateCohorts"></a>}}
\if{latex}{\out{\hypertarget{method-CohortTableHandler-insertOrUpdateCohorts}{}}}
\subsection{Method \code{insertOrUpdateCohorts()}}{
If there is no cohort with the same cohortId it is added to the cohortDefinitionSet,
If there is a cohort with the same cohortId, the cohort is updated in the cohortDefinitionSet
CohortDefinitionSet is generated and demographics is updated for only the cohorts that have changed
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortTableHandler$insertOrUpdateCohorts(cohortDefinitionSet)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cohortDefinitionSet}}{The cohort definition set to add.

deleteCohorts}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortTableHandler-deleteCohorts"></a>}}
\if{latex}{\out{\hypertarget{method-CohortTableHandler-deleteCohorts}{}}}
\subsection{Method \code{deleteCohorts()}}{
Deletes cohorts from the cohort table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortTableHandler$deleteCohorts(cohortIds)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cohortIds}}{The cohort ids to delete.

getCohortCounts}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortTableHandler-getCohortCounts"></a>}}
\if{latex}{\out{\hypertarget{method-CohortTableHandler-getCohortCounts}{}}}
\subsection{Method \code{getCohortCounts()}}{
Retrieves cohort counts from the cohort table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortTableHandler$getCohortCounts()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A tibble containing the cohort counts with names.

getCohortsSummary
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortTableHandler-getCohortsSummary"></a>}}
\if{latex}{\out{\hypertarget{method-CohortTableHandler-getCohortsSummary}{}}}
\subsection{Method \code{getCohortsSummary()}}{
Retrieves the summary of cohorts including cohort start and end year histograms and sex counts.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortTableHandler$getCohortsSummary()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A tibble containing cohort summary.

getCohortIdAndNames
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortTableHandler-getCohortIdAndNames"></a>}}
\if{latex}{\out{\hypertarget{method-CohortTableHandler-getCohortIdAndNames}{}}}
\subsection{Method \code{getCohortIdAndNames()}}{
Retrieves the cohort names.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortTableHandler$getCohortIdAndNames()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A vector with the name of the cohorts
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortTableHandler-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CohortTableHandler-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortTableHandler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
