% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureExtraction_covariateData.R
\name{covariateData_ATCgroups}
\alias{covariateData_ATCgroups}
\title{covariateData_ATCgroups}
\usage{
covariateData_ATCgroups(
  temporalStartDays = -99999,
  temporalEndDays = 99999,
  continuous = FALSE
)
}
\arguments{
\item{temporalStartDays}{Start day relative to index (-99999 by default)}

\item{temporalEndDays}{End day relative to index (99999 by default)}

\item{continuous}{Logical. If TRUE, the covariate data is continuous.}
}
\value{
A covariate settings object for ATC drug groups
}
\description{
Creates covariate settings for ATC drug groups
}
