% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureExtraction_covariateData.R
\name{covariateData_ATCgroups}
\alias{covariateData_ATCgroups}
\title{covariateData_ATCgroups}
\usage{
covariateData_ATCgroups(temporalStartDays = -99999, temporalEndDays = 99999)
}
\description{
covariateData_ATCgroups
}
