% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortGenerator_subsets.R
\name{createMatchingSubset}
\alias{createMatchingSubset}
\title{A definition of subset functions to be applied to a set of cohorts}
\usage{
createMatchingSubset(
  name = NULL,
  matchToCohortId,
  matchRatio = 10,
  matchSex = TRUE,
  matchBirthYear = TRUE,
  matchCohortStartDateWithInDuration = FALSE,
  newCohortStartDate = "keep",
  newCohortEndDate = "keep"
)
}
\arguments{
\item{name}{optional name of operator}

\item{matchToCohortId}{integer - cohort ids to match to}

\item{matchRatio}{matching ratio}

\item{matchSex}{match to target sex}

\item{matchBirthYear}{match to target birth year}

\item{matchCohortStartDateWithInDuration}{cohort_start_date of the matching subject must be within the duration of the target cohort}

\item{newCohortStartDate}{change cohort start date to the matched person}

\item{newCohortEndDate}{change cohort end date to the matched person}
}
\value{
a MatchingSubsetOperator instance
}
\description{
A definition of subset functions to be applied to a set of cohorts
}
