% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortDiagnostics.R
\name{.addFisherTestToCodeCounts}
\alias{.addFisherTestToCodeCounts}
\title{addFisherTestToCodeCounts}
\usage{
.addFisherTestToCodeCounts(timeCodeWasCounts)
}
\arguments{
\item{case_controls_counts}{table obtained from \code{getCodeCounts}.}
}
\value{
inputed table with appened columsn :
\itemize{
\item p : for the p-value
\item OR; for the ods ratio
}
}
\description{
performs an Fisher test in each row of the case_controls_counts table
}
