% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable_cohortsSummary.R
\name{.render_apex_plot}
\alias{.render_apex_plot}
\title{Renders an Apex chart for the given data.}
\usage{
.render_apex_plot(data)
}
\arguments{
\item{data}{A data frame containing the data to be plotted.}

\item{colorTimeHist}{The color for the time histogram (default is "#00BFFF").}
}
\value{
An Apex chart object.
}
\description{
Renders an Apex chart for the given data.
}
