% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeCodeWAS.R
\name{executeTimeCodeWAS}
\alias{executeTimeCodeWAS}
\title{CohortDiagnostics_runTimeCodeWAS}
\usage{
executeTimeCodeWAS(
  exportFolder,
  cohortTableHandler = NULL,
  cohortIdCases,
  cohortIdControls,
  analysisIds,
  temporalStartDays,
  temporalEndDays,
  minCellCount = 1,
  cohortDefinitionSet = NULL,
  databaseId = NULL,
  databaseName = NULL,
  databaseDescription = NULL,
  connectionDetails = NULL,
  connection = NULL,
  cdmDatabaseSchema = NULL,
  vocabularyDatabaseSchema = cdmDatabaseSchema,
  cohortDatabaseSchema = NULL,
  cohortTable = "cohort",
  vocabularyVersionCdm = NULL,
  vocabularyVersion = NULL
)
}
\arguments{
\item{exportFolder}{A character string specifying the folder path where the results will be exported.}

\item{cohortTableHandler}{An object of class 'CohortTableHandler' representing the cohort table handler.}

\item{cohortIdCases}{The ID of the cohort representing cases.}

\item{cohortIdControls}{The ID of the cohort representing controls.}

\item{analysisIds}{Numeric vector specifying the IDs of the analyses to be performed.}

\item{temporalStartDays}{Numeric vector specifying the start days for each time window.}

\item{temporalEndDays}{Numeric vector specifying the end days for each time window.}

\item{minCellCount}{An integer specifying the minimum cell count for Fisher's exact test.}
}
\value{
A logical value indicating if the function was executed successfully.
}
\description{
This function calculates the number of subjects with observation case and controls in each time window for a given cohort. It returns a data frame with the counts of cases and controls for each time window and covariate, as well as the results of a Fisher's exact test comparing the counts of cases and controls.
}
