% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureExtraction_covariateData.R
\name{YearOfBirth}
\alias{YearOfBirth}
\title{YearOfBirth}
\usage{
YearOfBirth(
  connection,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  cdmDatabaseSchema,
  cdmVersion = "5",
  cohortTable = "#cohort_person",
  cohortIds = c(-1),
  rowIdField = "subject_id",
  covariateSettings,
  aggregated = FALSE,
  minCharacterizationMean = 0
)
}
\description{
YearOfBirth
}
